Expression_Summ_Exp

+++ modules/local/salmon/quant_bam/main.nf
@@ -8,54 +8,42 @@ process SALMON_QUANT_BAM {
         'biocontainers/salmon:1.10.1--h7e5ed60_0' }"
 
     input:
-    tuple val(meta), path(bam)
-    path  transcript_fasta
+    tuple val(meta), path(reads)
+    path  index
     val   lib_type
 
     output:
-    tuple val(meta), path("${meta.id}") , emit: results
-    tuple val(meta), path("*info.json") , emit: json_info
-    path  "versions.yml"                , emit: versions
+    tuple val(meta), path("*_quant.sf")             , emit: quant
+    tuple val(meta), path("*info.json")             , emit: json_info, optional: true
+    tuple val(meta), path("*lib_format_counts.json"), emit: lib_format_counts, optional: true
+    path  "versions.yml"                            , emit: versions
 
     when:
     task.ext.when == null || task.ext.when
 
     script:
-    def args  = task.ext.args ?: ''
-    def prefix          = task.ext.prefix ?: "${meta.id}"
-    def reference       = "-t $transcript_fasta"
-    def input_reads     = "-a $bam"
+    prefix   = task.ext.prefix ?: "${meta.id}"
 
-    def strandedness_opts = [
-        'A',  'U',   'SF',  'SR',
-        'IS', 'IU' , 'ISF', 'ISR',
-        'OS', 'OU' , 'OSF', 'OSR',
-        'MS', 'MU' , 'MSF', 'MSR'
-    ]
-    def strandedness =  'A'
-    if (lib_type) {
-        if (strandedness_opts.contains(lib_type)) {
-            strandedness = lib_type
-        } else {
-            log.info "[Salmon Quant] Invalid library type specified '--libType=${lib_type}', defaulting to auto-detection with '--libType=A'."
-        }
-    } else {
-        strandedness = meta.single_end ? 'U' : 'IU'
-        if (meta.strandedness == 'forward') {
-            strandedness = meta.single_end ? 'SF' : 'ISF'
-        } else if (meta.strandedness == 'reverse') {
-            strandedness = meta.single_end ? 'SR' : 'ISR'
-        }
-    }
+    def reads1 = [], reads2 = []
+    meta.single_end ? [reads].flatten().each{reads1 << it} : reads.eachWithIndex{ v, ix -> ( ix & 1 ? reads2 : reads1) << v }
     """
     salmon quant \\
-        $reference \\
-        --libType=$strandedness \\
-        $input_reads \\
-        $args \\
-        -o $prefix
+        -i $index \\
+        -o $prefix \\
+        -l $lib_type \\
+        -p ${task.cpus} \\
+        --validateMappings \\
+        -1 ${reads1.join(",")} -2 ${reads2.join(",")} \\
 
-    cp $prefix/aux_info/meta_info.json "${prefix}_meta_info.json"
+    if [ -f $prefix/aux_info/meta_info.json ]; then
+        cp $prefix/aux_info/meta_info.json "${prefix}_meta_info.json"
+    fi
+    if [ -f $prefix/lib_format_counts.json ]; then
+        cp $prefix/lib_format_counts.json "${prefix}_lib_format_counts.json"
+    fi
+    if [ -f $prefix/quant.sf ]; then
+        cp $prefix/quant.sf "${prefix}_quant.sf"
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
@@ -64,10 +52,11 @@ process SALMON_QUANT_BAM {
     """
 
     stub:
-    def prefix = task.ext.prefix ?: "${meta.id}"
+    prefix = task.ext.prefix ?: "${meta.id}"
     """
     mkdir ${prefix}
     touch ${prefix}_meta_info.json
+    touch ${prefix}_lib_format_counts.json
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
